cleaning up files
