+++ src/main/java/com/stolensugar/web/voiceCommands/Mochi/Cards.java
@@ -1,4 +1,4 @@
-package com.stolensugar.web.controller.mappers.Mochi;
+package com.stolensugar.web.voiceCommands.Mochi;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,7 +6,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.*;
 
-import javax.annotation.processing.Generated;
 import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -18,9 +17,8 @@ import java.util.List;
 @NoArgsConstructor(force = true, access = AccessLevel.PUBLIC)
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown=true)
-public class Templates {
+public class Cards {
 
     @JsonProperty("~#list")
-    private @Getter @Setter List<SingleTemplate> list = null;
-
+    private @Getter @Setter List<SingleCard> list = null;
 }
