+++ src/main/java/com/stolensugar/web/voiceCommands/Mochi/FieldParams.java
@@ -1,25 +1,22 @@
-package com.stolensugar.web.controller.mappers.Mochi;
+package com.stolensugar.web.voiceCommands.Mochi;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.*;
 
-import javax.annotation.processing.Generated;
-import java.util.List;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "~#list"
+        "~:id",
+        "~:value"
 })
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @ToString
 @NoArgsConstructor(force = true, access = AccessLevel.PUBLIC)
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown=true)
-public class Cards {
-
-    @JsonProperty("~#list")
-    private @Getter @Setter List<SingleCard> list = null;
+public class FieldParams {
+    @JsonProperty("~:id")
+    private @Getter @Setter String id;
+    @JsonProperty("~:value")
+    private @Getter @Setter String value;
 }
