+++ src/main/java/com/stolensugar/web/voiceCommands/Mochi/SingleCard.java
@@ -1,4 +1,4 @@
-package com.stolensugar.web.controller.mappers.Mochi;
+package com.stolensugar.web.voiceCommands.Mochi;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
