+++ src/main/java/com/stolensugar/web/voiceCommands/Mochi/Templates.java
@@ -1,4 +1,4 @@
-package com.stolensugar.web.controller.mappers.Mochi;
+package com.stolensugar.web.voiceCommands.Mochi;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,28 +6,20 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.*;
 
-import javax.annotation.processing.Generated;
-import java.util.ArrayList;
 import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "~:decks",
-        "~:version",
-        "~:templates"
+        "~#list"
 })
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @ToString
 @NoArgsConstructor(force = true, access = AccessLevel.PUBLIC)
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown=true)
-public class MochiDeck {
+public class Templates {
 
-    @JsonProperty("~:decks")
-    private @Getter @Setter List<Deck> decks = new ArrayList<>();
-    @JsonProperty("~:version")
-    private @Getter @Setter Integer version = 2;
-    @JsonProperty("~:templates")
-    private @Getter @Setter Templates templates;
+    @JsonProperty("~#list")
+    private @Getter @Setter List<SingleTemplate> list = null;
 
 }
